build: Bump meson dependency to 1.7.0
Later commits introduce a meson.options file for configuring the build.
Previously, this file was called meson_options.txt but in Meson 1.1.0
that was changed to the new one. To avoid introducing a deprecated name
that just will have to be renamed later, bump the minimum Meson version
required now. 1.7.0 was chosen as it's the version that will be present
in the next Debian stable release. Additionally, the latest Alpine
stable at the time of writing has Meson 1.8.1, so this works for
postmarketOS too. While it could've been bumped to 1.1.0 instead, 1.7.0
means we can wait longer until we have to bump this again.

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/219>