tweaks-preference-page: Fix file_picker_label life cycle
Appending to a box sinks the ref so we can't sink or unref it
later without taking a reference on our own first. But as the label
won't outlive the containing widget there's no need for us to
keep a reference anyway.

Should we ever intend to use the metadata longer than the generated
widget we would sink the ref in `setting_data_to_file_widget` and
would unref on dispose.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/269>