conf-tweaks: Remove nonsensical assertion
I'm not sure why I added this assertion, but my guess would be that
either `from` also used to be a pointer, or I just didn't really think
about that it wasn't one when adding that. Either way, it makes no sense
since it's perfectly valid for the weigh of a setting to be 0. Asserting
that `into` is not 0 makes sense since it is a pointer and this function
will crash if `into` is a NULL pointer.

Closes https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/issues/88

Fixes b113b068e26ce11cb645b5570d9f00a999cb8de0

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/271>