conf-tweaks: Only attempt to create new Xresources on G_FILE_ERROR_NOENT
If there was some other error, we probably don't want to try to make a
new file since something is weird with the location we're trying to
write to.

Additionally, shuffle around some code to make it more readable and
clearer what's going on. It seems unexpected to me that
write_new_xresources () expects an error that already is filled and then
clears it. Move this to the caller.

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/278>